backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
cryptography==41.0.7
Flask-Mail==0.9.1
